fix(github-action): Properly handle configuration (and auth) seeding
This commit fixes an issue whereby configuration was not correctly
seeded into the "main" system.  All initialization, including
Unikraft Cloud authentication detection, is now handled in `main()`
and overrides via env vars + CLI are handled in the `Run` portion
of the command.  This ensures that access to config and
authentication is consistent across all subsystems.

Signed-off-by: Alexander Jung <alex@unikraft.io>