fix(pkg): Prioritize kernel packaging
This commit re-orders the kernel packager to be first such that
it can be used in the context of an existing working directory.
Since the kernel is a known flag, we can safely assume that
the user knows what they are doing and can allow them to
manipulate the package manager directly.

Signed-off-by: Alexander Jung <alex@unikraft.io>