feat(manifest): Introduce options to customize default channel name
This commit introduces two options (one for the manager and one for
the individual manifest provider) which allows for customizing the
default channel name.  Before, this was hard coded to `stable`.
Now this is exposed as a public constant for the default name of
"stable" but can be set arbitrarily through relevant options.

Signed-off-by: Alexander Jung <alex@unikraft.io>