refactor(manifest)!: Introduce allow pull by channel and version
This commit reworks the manifest provider interface and removes
`PullManifest` in favor `PullChannel` and `PullVersion`.  This
bubbles up the choice to the package's `Pull` method which
ultimately used it but provides a more flexible public interface
for interacting with a manifest and its channels and versions
and gaining access to individual resources.

Signed-off-by: Alexander Jung <alex@unikraft.io>