refactor(oci)!: Add `handle` to `NewPackageFromTarget` signature
Previously the `NewPackageFromTarget` would attempt to seed this
by looking at KraftKit's configuration.  Instead, pass this
directly as a required argument.  Since this is only used in one
place, it is a small refactor.  However, this ultimately allows
the for the OCI package manager to be instantiated with its
manager options, i.e. `WithDirectory` or `WithContainerd` and
thus the handler can be used during packaging without having to
re-read configuration.

Signed-off-by: Alexander Jung <alex@unikraft.io>