fix(app): Prioritize project targets over template targets
Do not merge template targets.  If a user specifies a template
in their `Kraftfile` and supply a list of `targets`, they expect
there to only be those targets in _their_ specified list.
Adopting the targets from the template would mean that
programmatically we'd have to search for our target or specify
a name, which makes the process more cumbersome.

Signed-off-by: Alexander Jung <alex@unikraft.io>