feat(github-action): Support setting custom list of source manifests
In order to make this possible, a new actions variable, `manifests`,
is exposed.  This is a YAML list.  Authentication for this is also
possible but we must also bump to the latest version of KraftKit.
Re-organization of the GitHub action's main sequence is also
necessary in order to properly seed the manifest provider's
authentication for such manifest sources.

Signed-off-by: Alexander Jung <alex@unikraft.io>