fix(log): Better handling of EOF from instance logs
This commit fixes an issue when tailing logs from an exited instance
by using the updated UKC Go SDK which now incorporates better
signalling of EOF.  At the same time, this rework prevents having
to check the status with a delay of 1 second which results in the
"The instance has exited" message to be returned immediately.  This
is because the status is first checked before any tailing occurs
with additional logic to now check for the EOF error.

Signed-off-by: Alexander Jung <alex@unikraft.io>